pyproject.toml
[project]
name = "python-aoc-template"
version = "0.1.0"
description = "Advent of Code Python Template with automated boilerplate"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pytest>=7.4.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "beautifulsoup4>=4.12.0",
]

[project.scripts]
aoc = "aoc_template.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["aoc_template"]

[tool.pytest.ini_options]
testpaths = ["day*/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
]
